week-1-chatbots/03-pdf-resume-reviewer/requirements.txt
streamlit 
langchain 
langchain-openai 
python-dotenv 
pymupdf
